Make ServerChatEvent extend MessageEvent
